refactor(scan): ensure we use the parsed election
Apparently TS favors the `.json` file when importing, so the `election.ts` files were being ignored. Go figure!
