+++ apps/module-scan/src/backup.test.ts
@@ -5,7 +5,7 @@ import { Database } from 'sqlite3'
 import { fileSync } from 'tmp'
 import { Entry, fromBuffer, ZipFile } from 'yauzl'
 import ZipStream from 'zip-stream'
-import election from '../test/fixtures/2020-choctaw/election'
+import { election } from '../test/fixtures/2020-choctaw'
 import backup, { Backup } from './backup'
 import Store from './store'
 import { asElectionDefinition } from '@votingworks/fixtures'
