+++ apps/module-scan/src/interpreter.test.ts
@@ -3,10 +3,10 @@ import { metadataFromBytes } from '@votingworks/hmpb-interpreter'
 import { AdjudicationReason, Election } from '@votingworks/types'
 import { readFile } from 'fs-extra'
 import { join } from 'path'
-import choctaw2020Election from '../test/fixtures/2020-choctaw/election'
+import { election as choctaw2020Election } from '../test/fixtures/2020-choctaw'
 import * as general2020Fixtures from '../test/fixtures/2020-general'
 import * as choctaw2020SpecialFixtures from '../test/fixtures/choctaw-2020-09-22-f30480cc99'
-import stateOfHamiltonElection from '../test/fixtures/state-of-hamilton/election'
+import { election as stateOfHamiltonElection } from '../test/fixtures/state-of-hamilton'
 import Interpreter, {
   BlankPage,
   getBallotImageData,
