+++ apps/module-scan/src/server.test.ts
@@ -21,7 +21,7 @@ import request from 'supertest'
 import { dirSync } from 'tmp'
 import { mocked } from 'ts-jest/dist/utils/testing'
 import { v4 as uuid } from 'uuid'
-import election from '../test/fixtures/state-of-hamilton/election'
+import { election } from '../test/fixtures/state-of-hamilton'
 import zeroRect from '../test/fixtures/zeroRect'
 import { makeMock } from '../test/util/mocks'
 import Importer from './importer'
