+++ apps/module-scan/src/store.test.ts
@@ -8,7 +8,7 @@ import {
 import { promises as fs } from 'fs'
 import * as tmp from 'tmp'
 import { v4 as uuid } from 'uuid'
-import election from '../test/fixtures/state-of-hamilton/election'
+import { election } from '../test/fixtures/state-of-hamilton'
 import zeroRect from '../test/fixtures/zeroRect'
 import Store from './store'
 import {
