--- apps/module-scan/test/fixtures/choctaw-2020-09-22-f30480cc99/election.ts
@@ -1,4 +1,0 @@
-import { parseElection } from '@votingworks/types'
-import electionJSON from './election.json'
-
-export default parseElection(electionJSON)
