+++ apps/module-scan/test/fixtures/choctaw-2020-09-22-f30480cc99/index.ts
@@ -1,7 +1,8 @@
+import { parseElection } from '@votingworks/types'
 import { join } from 'path'
-import election from './election'
+import electionJSON from './election.json'
 
-export { election }
+export const election = parseElection(electionJSON)
 export const ballotPdf = join(__dirname, 'ballot.pdf')
 export const ballot6522Pdf = join(__dirname, 'ballot-6522.pdf')
 export const blankPage1 = join(__dirname, 'blank-p1.png')
