+++ apps/module-scan/test/fixtures/choctaw-mock-general-election-2020/index.ts
@@ -1,9 +1,10 @@
+import { parseElection } from '@votingworks/types'
 import { readFileSync } from 'fs-extra'
 import { join } from 'path'
 import { BallotPackageManifest } from '../../../src/types'
-import election from './election'
+import electionJSON from './election.json'
 
-export { election }
+export const election = parseElection(electionJSON)
 export const root = __dirname
 export const manifest: BallotPackageManifest = JSON.parse(
   readFileSync(join(__dirname, 'manifest.json'), 'utf8')
