--- apps/module-scan/test/fixtures/state-of-hamilton/election.ts
@@ -1,4 +1,0 @@
-import { parseElection } from '@votingworks/types'
-import electionJSON from './election.json'
-
-export default parseElection(electionJSON)
