+++ libs/hmpb-interpreter/test/fixtures/choctaw-2020-09-22-f30480cc99/index.ts
@@ -1,8 +1,9 @@
+import { parseElection } from '@votingworks/types'
 import { join } from 'path'
 import { Fixture } from '../../fixtures'
-import election from './election'
+import electionJSON from './election.json'
 
-export { election }
+export const election = parseElection(electionJSON)
 export const ballotPdf = new Fixture(join(__dirname, 'ballot.pdf'))
 export const blankPage1 = new Fixture(join(__dirname, 'blank-p1.png'))
 export const blankPage2 = new Fixture(join(__dirname, 'blank-p2.png'))
