--- libs/hmpb-interpreter/test/fixtures/choctaw-county-2020-general-election/election.ts
@@ -1,4 +1,0 @@
-import { parseElection } from '@votingworks/types'
-import electionJSON from './election.json'
-
-export default parseElection(electionJSON)
