+++ libs/hmpb-interpreter/test/fixtures/choctaw-county-2020-general-election/index.ts
@@ -1,6 +1,9 @@
+import { parseElection } from '@votingworks/types'
 import { join } from 'path'
 import { Fixture } from '../../fixtures'
+import electionJSON from './election.json'
 
+export const election = parseElection(electionJSON)
 export const filledInPage1_01 = new Fixture(
   join(__dirname, 'filled-in-p1-01.png')
 )
@@ -56,5 +59,3 @@ export const eastWeirBlankPage1 = new Fixture(
 export const eastWeirBlankPage2 = new Fixture(
   join(__dirname, 'ballot-east-weir-p2.png')
 )
-
-export { default as election } from './election'
