--- libs/hmpb-interpreter/test/fixtures/choctaw-county-mock-general-election-choctaw-2020-e87f23ca2c/election.ts
@@ -1,4 +1,0 @@
-import { parseElection } from '@votingworks/types'
-import electionJSON from './election.json'
-
-export default parseElection(electionJSON)
