+++ libs/hmpb-interpreter/test/fixtures/choctaw-county-mock-general-election-choctaw-2020-e87f23ca2c/index.ts
@@ -1,10 +1,10 @@
+import { parseElection } from '@votingworks/types'
 import { join } from 'path'
 import { Fixture } from '../../fixtures'
-import election from './election'
+import electionJSON from './election.json'
 
+export const election = parseElection(electionJSON)
 export const blankPage1 = new Fixture(join(__dirname, 'blank-p1.png'))
 export const blankPage2 = new Fixture(join(__dirname, 'blank-p2.png'))
 export const filledInPage1 = new Fixture(join(__dirname, 'filled-in-p1.png'))
 export const filledInPage2 = new Fixture(join(__dirname, 'filled-in-p2.png'))
-
-export { election }
