--- libs/hmpb-interpreter/test/fixtures/election-4e31cb17d8-ballot-style-77-precinct-oaklawn-branch-library/election.ts
@@ -1,6 +1,0 @@
-import electionJSON from './election.json'
-import { parseElection } from '@votingworks/types'
-
-const election = parseElection(electionJSON)
-
-export default election
