+++ libs/hmpb-interpreter/test/fixtures/election-4e31cb17d8-ballot-style-77-precinct-oaklawn-branch-library/index.ts
@@ -1,7 +1,9 @@
+import { parseElection } from '@votingworks/types'
 import { join } from 'path'
 import { Fixture } from '../../fixtures'
-import election from './election'
+import electionJSON from './election.json'
 
+export const election = parseElection(electionJSON)
 export const electionPath = join(__dirname, 'election.json')
 export const blankPage1 = new Fixture(join(__dirname, 'blank-p1.jpg'))
 export const blankPage2 = new Fixture(join(__dirname, 'blank-p2.jpg'))
@@ -10,5 +12,3 @@ export const filledInPage2 = new Fixture(join(__dirname, 'filled-in-p2.jpg'))
 export const partialBorderPage2 = new Fixture(
   join(__dirname, 'extra-contest-detected-p2.jpg')
 )
-
-export { election }
