--- libs/hmpb-interpreter/test/fixtures/election-5c6e578acf-state-of-hamilton-2020/election.ts
@@ -1,4 +1,0 @@
-import { parseElection } from '@votingworks/types'
-import electionJSON from './election.json'
-
-export default parseElection(electionJSON)
