+++ libs/hmpb-interpreter/test/fixtures/election-5c6e578acf-state-of-hamilton-2020/index.ts
@@ -1,7 +1,9 @@
+import { parseElection } from '@votingworks/types'
 import { join } from 'path'
 import { Fixture } from '../../fixtures'
-import election from './election'
+import electionJSON from './election.json'
 
+export const election = parseElection(electionJSON)
 export const blankPage1 = new Fixture(join(__dirname, 'blank-p1.jpg'))
 export const blankPage2 = new Fixture(join(__dirname, 'blank-p2.jpg'))
 export const blankPage3 = new Fixture(join(__dirname, 'blank-p3.jpg'))
@@ -15,4 +17,3 @@ export const filledInPage5 = new Fixture(join(__dirname, 'filled-in-p5.jpg'))
 export const filledInPage5YesNoOvervote = new Fixture(
   join(__dirname, 'filled-in-p5-yesno-overvote.jpg')
 )
-export { election }
