--- libs/hmpb-interpreter/test/fixtures/election-7c61368c3b-choctaw-general-2020/election.ts
@@ -1,4 +1,0 @@
-import { parseElection } from '@votingworks/types'
-import electionJSON from './election.json'
-
-export default parseElection(electionJSON)
