+++ libs/hmpb-interpreter/test/fixtures/election-7c61368c3b-choctaw-general-2020/index.ts
@@ -1,9 +1,10 @@
+import { parseElection } from '@votingworks/types'
 import { join } from 'path'
 import { Fixture } from '../../fixtures'
-import election from './election'
+import electionJSON from './election.json'
 
+export const election = parseElection(electionJSON)
 export const blankPage1 = new Fixture(join(__dirname, 'blank-p1.png'))
 export const blankPage2 = new Fixture(join(__dirname, 'blank-p2.png'))
 export const filledInPage1 = new Fixture(join(__dirname, 'filled-in-p1.png'))
 export const filledInPage2 = new Fixture(join(__dirname, 'filled-in-p2.png'))
-export { election }
