--- libs/hmpb-interpreter/test/fixtures/election-98f5203139-choctaw-general-2019/election.ts
@@ -1,4 +1,0 @@
-import { parseElection } from '@votingworks/types'
-import electionJSON from './election.json'
-
-export default parseElection(electionJSON)
