+++ libs/hmpb-interpreter/test/fixtures/election-98f5203139-choctaw-general-2019/index.ts
@@ -1,7 +1,9 @@
+import { parseElection } from '@votingworks/types'
 import { join } from 'path'
 import { Fixture } from '../../fixtures'
-import election from './election'
+import electionJSON from './election.json'
 
+export const election = parseElection(electionJSON)
 export const blankPage1 = new Fixture(join(__dirname, 'blank-p1.jpg'))
 export const blankPage2 = new Fixture(join(__dirname, 'blank-p2.jpg'))
 export const blankPage3 = new Fixture(join(__dirname, 'blank-p3.jpg'))
@@ -9,4 +11,3 @@ export const filledInPage1 = new Fixture(join(__dirname, 'filled-in-p1.jpg'))
 export const filledInPage2 = new Fixture(join(__dirname, 'filled-in-p2.jpg'))
 export const borderPage1 = new Fixture(join(__dirname, 'border-p1.jpg'))
 export const borderPage3 = new Fixture(join(__dirname, 'border-p3.jpg'))
-export { election }
